添加EXPLOSION_DESTROY属性

+++ src/main/kotlin/com/atsuishio/superbwarfare/entity/projectile/FastThrowableProjectile.kt
@@ -74,6 +74,7 @@ abstract class FastThrowableProjectile : ThrowableItemProjectile, IFastMotionSyn
     protected var penetratingValue: Boolean = false
     protected val effectsValue: MutableSet<MobEffectInstance> = hashSetOf()
     protected var underwaterMotionScaleValue = 0.75f
+    protected var explosionDestroyValue = true
 
     override fun getDamage(): Float = damageValue
     override fun setDamage(value: Float) {
@@ -130,6 +131,11 @@ abstract class FastThrowableProjectile : ThrowableItemProjectile, IFastMotionSyn
         underwaterMotionScaleValue = value
     }
 
+    override fun hasExplosionDestroy(): Boolean = explosionDestroyValue
+    override fun setExplosionDestroy(value: Boolean) {
+        explosionDestroyValue = value
+    }
+
     private var isFastMoving = false
 
     var exploded: Boolean = false
@@ -181,6 +187,9 @@ abstract class FastThrowableProjectile : ThrowableItemProjectile, IFastMotionSyn
         if (compound.contains("UnderwaterMotionScale")) {
             this.underwaterMotionScaleValue = compound.getFloat("UnderwaterMotionScale")
         }
+        if (compound.contains("ExplosionDestroy")) {
+            this.explosionDestroyValue = compound.getBoolean("ExplosionDestroy")
+        }
 
         this.effectsValue.addAll(PotionUtils.getCustomEffects(compound))
     }
@@ -209,6 +218,7 @@ abstract class FastThrowableProjectile : ThrowableItemProjectile, IFastMotionSyn
         if (this.underwaterMotionScaleValue > 0) {
             compound.putFloat("UnderwaterMotionScale", this.underwaterMotionScaleValue)
         }
+        compound.putBoolean("ExplosionDestroy", this.explosionDestroyValue)
 
         if (!this.effectsValue.isEmpty()) {
             val list = ListTag()
@@ -519,21 +529,15 @@ abstract class FastThrowableProjectile : ThrowableItemProjectile, IFastMotionSyn
         val resultPos = blockHitResult.blockPos
         val hardness = this.level().getBlockState(resultPos).block.defaultDestroyTime()
         if (hardness != -1f) {
-            if (ExplosionConfig.EXPLOSION_DESTROY.get()) {
-                if (firstHit) {
-                    causeExplode(blockHitResult.location)
-                    firstHit = false
-                    queueServerWork(3) { this.discard() }
-                }
-                if (ExplosionConfig.EXTRA_EXPLOSION_EFFECT.get()) {
-                    this.level().destroyBlock(resultPos, true)
-                }
+            if (firstHit) {
+                causeExplode(blockHitResult.location)
+                firstHit = false
+                queueServerWork(3) { this.discard() }
+            }
+            if (ExplosionConfig.EXPLOSION_DESTROY.get() && ExplosionConfig.EXTRA_EXPLOSION_EFFECT.get() && this.explosionDestroyValue) {
+                this.level().destroyBlock(resultPos, true)
             }
         } else {
-            causeExplode(blockHitResult.location)
-            this.discard()
-        }
-        if (!ExplosionConfig.EXPLOSION_DESTROY.get()) {
             causeExplode(blockHitResult.location)
             this.discard()
         }
@@ -546,6 +550,7 @@ abstract class FastThrowableProjectile : ThrowableItemProjectile, IFastMotionSyn
             .radius(explosionRadiusValue)
             .position(vec3)
             .beast(this.isBeast())
+            .destroyBlock(explosionDestroyValue)
     }
 
     open fun causeRangedEffects(vec3: Vec3) {
